prov/psm2: Put trigger implementation into a separate file
Also update copyright information for files that have been updated
recently.
