+++ prov/psm2/src/psmx2.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2013-2017 Intel Corporation. All rights reserved.
+ * Copyright (c) 2013-2018 Intel Corporation. All rights reserved.
  * Copyright (c) 2016 Cisco Systems, Inc. All rights reserved.
  *
  * This software is available to you under a choice of one of two
@@ -491,235 +491,7 @@ struct psmx2_fid_cq {
 	uint8_t				error_data[PSMX2_ERR_DATA_SIZE];
 };
 
-enum psmx2_triggered_op {
-	PSMX2_TRIGGERED_SEND,
-	PSMX2_TRIGGERED_SENDV,
-	PSMX2_TRIGGERED_RECV,
-	PSMX2_TRIGGERED_TSEND,
-	PSMX2_TRIGGERED_TSENDV,
-	PSMX2_TRIGGERED_TRECV,
-	PSMX2_TRIGGERED_WRITE,
-	PSMX2_TRIGGERED_WRITEV,
-	PSMX2_TRIGGERED_READ,
-	PSMX2_TRIGGERED_READV,
-	PSMX2_TRIGGERED_ATOMIC_WRITE,
-	PSMX2_TRIGGERED_ATOMIC_WRITEV,
-	PSMX2_TRIGGERED_ATOMIC_READWRITE,
-	PSMX2_TRIGGERED_ATOMIC_READWRITEV,
-	PSMX2_TRIGGERED_ATOMIC_COMPWRITE,
-	PSMX2_TRIGGERED_ATOMIC_COMPWRITEV,
-};
-
-struct psmx2_trigger {
-	enum psmx2_triggered_op	op;
-	struct psmx2_fid_cntr	*cntr;
-	size_t			threshold;
-	union {
-		struct {
-			struct fid_ep	*ep;
-			const void	*buf;
-			size_t		len;
-			void		*desc;
-			fi_addr_t	dest_addr;
-			void		*context;
-			uint64_t	flags;
-			uint64_t	data;
-		} send;
-		struct {
-			struct fid_ep	*ep;
-			const struct iovec *iov;
-			size_t		count;
-			void		**desc;
-			fi_addr_t	dest_addr;
-			void		*context;
-			uint64_t	flags;
-			uint64_t	data;
-		} sendv;
-		struct {
-			struct fid_ep	*ep;
-			void		*buf;
-			size_t		len;
-			void		*desc;
-			fi_addr_t	src_addr;
-			void		*context;
-			uint64_t	flags;
-		} recv;
-		struct {
-			struct fid_ep	*ep;
-			const void	*buf;
-			size_t		len;
-			void		*desc;
-			fi_addr_t	dest_addr;
-			uint64_t	tag;
-			void		*context;
-			uint64_t	flags;
-			uint64_t	data;
-		} tsend;
-		struct {
-			struct fid_ep	*ep;
-			const struct iovec *iov;
-			size_t		count;
-			void		**desc;
-			fi_addr_t	dest_addr;
-			uint64_t	tag;
-			void		*context;
-			uint64_t	flags;
-			uint64_t	data;
-		} tsendv;
-		struct {
-			struct fid_ep	*ep;
-			void		*buf;
-			size_t		len;
-			void		*desc;
-			fi_addr_t	src_addr;
-			uint64_t	tag;
-			uint64_t	ignore;
-			void		*context;
-			uint64_t	flags;
-		} trecv;
-		struct {
-			struct fid_ep	*ep;
-			const void	*buf;
-			size_t		len;
-			void		*desc;
-			fi_addr_t	dest_addr;
-			uint64_t	addr;
-			uint64_t	key;
-			void		*context;
-			uint64_t	flags;
-			uint64_t	data;
-		} write;
-		struct {
-			struct fid_ep	*ep;
-			const struct iovec *iov;
-			size_t		count;
-			void		*desc;
-			fi_addr_t	dest_addr;
-			uint64_t	addr;
-			uint64_t	key;
-			void		*context;
-			uint64_t	flags;
-			uint64_t	data;
-		} writev;
-		struct {
-			struct fid_ep	*ep;
-			void		*buf;
-			size_t		len;
-			void		*desc;
-			fi_addr_t	src_addr;
-			uint64_t	addr;
-			uint64_t	key;
-			void		*context;
-			uint64_t	flags;
-		} read;
-		struct {
-			struct fid_ep	*ep;
-			const struct iovec *iov;
-			size_t		count;
-			void		*desc;
-			fi_addr_t	src_addr;
-			uint64_t	addr;
-			uint64_t	key;
-			void		*context;
-			uint64_t	flags;
-		} readv;
-		struct {
-			struct fid_ep	*ep;
-			const void	*buf;
-			size_t		count;
-			void		*desc;
-			fi_addr_t	dest_addr;
-			uint64_t	addr;
-			uint64_t	key;
-			enum fi_datatype datatype;
-			enum fi_op	atomic_op;
-			void		*context;
-			uint64_t	flags;
-		} atomic_write;
-		struct {
-			struct fid_ep	*ep;
-			const struct fi_ioc *iov;
-			size_t		count;
-			void		*desc;
-			fi_addr_t	dest_addr;
-			uint64_t	addr;
-			uint64_t	key;
-			enum fi_datatype datatype;
-			enum fi_op	atomic_op;
-			void		*context;
-			uint64_t	flags;
-		} atomic_writev;
-		struct {
-			struct fid_ep	*ep;
-			const void	*buf;
-			size_t		count;
-			void		*desc;
-			void		*result;
-			void		*result_desc;
-			fi_addr_t	dest_addr;
-			uint64_t	addr;
-			uint64_t	key;
-			enum fi_datatype datatype;
-			enum fi_op	atomic_op;
-			void		*context;
-			uint64_t	flags;
-		} atomic_readwrite;
-		struct {
-			struct fid_ep	*ep;
-			const struct fi_ioc *iov;
-			size_t		count;
-			void		**desc;
-			struct fi_ioc	*resultv;
-			void		**result_desc;
-			size_t		result_count;
-			fi_addr_t	dest_addr;
-			uint64_t	addr;
-			uint64_t	key;
-			enum fi_datatype datatype;
-			enum fi_op	atomic_op;
-			void		*context;
-			uint64_t	flags;
-		} atomic_readwritev;
-		struct {
-			struct fid_ep	*ep;
-			const void	*buf;
-			size_t		count;
-			void		*desc;
-			const void	*compare;
-			void		*compare_desc;
-			void		*result;
-			void		*result_desc;
-			fi_addr_t	dest_addr;
-			uint64_t	addr;
-			uint64_t	key;
-			enum fi_datatype datatype;
-			enum fi_op	atomic_op;
-			void		*context;
-			uint64_t	flags;
-		} atomic_compwrite;
-		struct {
-			struct fid_ep	*ep;
-			const struct fi_ioc *iov;
-			size_t		count;
-			void		**desc;
-			const struct fi_ioc *comparev;
-			void		**compare_desc;
-			size_t		compare_count;
-			struct fi_ioc	*resultv;
-			void		**result_desc;
-			size_t		result_count;
-			fi_addr_t	dest_addr;
-			uint64_t	addr;
-			uint64_t	key;
-			enum fi_datatype datatype;
-			enum fi_op	atomic_op;
-			void		*context;
-			uint64_t	flags;
-		} atomic_compwritev;
-	};
-	struct psmx2_trigger *next;	/* used for randomly accessed trigger list */
-	struct slist_entry list_entry;	/* used for ready-to-fire trigger queue */
-};
+struct psmx2_trigger;
 
 struct psmx2_fid_cntr {
 	union {
@@ -1049,8 +821,6 @@ int	psmx2_am_process_send(struct psmx2_trx_ctxt *trx_ctxt,
 				struct psmx2_am_request *req);
 int	psmx2_am_process_rma(struct psmx2_trx_ctxt *trx_ctxt,
 				struct psmx2_am_request *req);
-int	psmx2_process_trigger(struct psmx2_trx_ctxt *trx_ctxt,
-				struct psmx2_trigger *trigger);
 int	psmx2_am_rma_handler(psm2_am_token_t token, psm2_amarg_t *args,
 			     int nargs, void *src, uint32_t len,
 			     void *hctx);
@@ -1148,112 +918,6 @@ static inline void psmx2_progress_all(struct psmx2_fid_domain *domain)
 	}
 	psmx2_unlock(&domain->trx_ctxt_lock, 1);
 }
-
-/* The following functions are used by triggered operations */
-
-ssize_t psmx2_send_generic(
-			struct fid_ep *ep,
-			const void *buf, size_t len,
-			void *desc, fi_addr_t dest_addr,
-			void *context, uint64_t flags,
-			uint64_t data);
-
-ssize_t psmx2_sendv_generic(
-			struct fid_ep *ep,
-			const struct iovec *iov, void *desc,
-			size_t count, fi_addr_t dest_addr,
-			void *context, uint64_t flags,
-			uint64_t data);
-
-ssize_t psmx2_recv_generic(
-			struct fid_ep *ep,
-			void *buf, size_t len, void *desc,
-			fi_addr_t src_addr, void *context,
-			uint64_t flags);
-
-ssize_t psmx2_tagged_send_generic(
-			struct fid_ep *ep,
-			const void *buf, size_t len,
-			void *desc, fi_addr_t dest_addr,
-			uint64_t tag, void *context,
-			uint64_t flags, uint64_t data);
-
-ssize_t psmx2_tagged_sendv_generic(
-			struct fid_ep *ep,
-			const struct iovec *iov, void *desc,
-			size_t count, fi_addr_t dest_addr,
-			uint64_t tag, void *context,
-			uint64_t flags, uint64_t data);
-
-ssize_t psmx2_tagged_recv_generic(
-			struct fid_ep *ep,
-			void *buf, size_t len,
-			void *desc, fi_addr_t src_addr,
-			uint64_t tag, uint64_t ignore,
-			void *context, uint64_t flags);
-
-ssize_t psmx2_write_generic(
-			struct fid_ep *ep,
-			const void *buf, size_t len,
-			void *desc, fi_addr_t dest_addr,
-			uint64_t addr, uint64_t key,
-			void *context, uint64_t flags,
-			uint64_t data);
-
-ssize_t psmx2_writev_generic(
-			struct fid_ep *ep,
-			const struct iovec *iov, void **desc,
-			size_t count, fi_addr_t dest_addr,
-			uint64_t addr, uint64_t key,
-			void *context, uint64_t flags,
-			uint64_t data);
-
-ssize_t psmx2_read_generic(
-			struct fid_ep *ep,
-			void *buf, size_t len,
-			void *desc, fi_addr_t src_addr,
-			uint64_t addr, uint64_t key,
-			void *context, uint64_t flags);
-
-ssize_t psmx2_readv_generic(
-			struct fid_ep *ep,
-			const struct iovec *iov, void *desc,
-			size_t count, fi_addr_t src_addr,
-			uint64_t addr, uint64_t key,
-			void *context, uint64_t flags);
-
-ssize_t psmx2_atomic_write_generic(
-			struct fid_ep *ep,
-			const void *buf,
-			size_t count, void *desc,
-			fi_addr_t dest_addr,
-			uint64_t addr, uint64_t key,
-			enum fi_datatype datatype,
-			enum fi_op op, void *context,
-			uint64_t flags);
-
-ssize_t psmx2_atomic_readwrite_generic(
-			struct fid_ep *ep,
-			const void *buf,
-			size_t count, void *desc,
-			void *result, void *result_desc,
-			fi_addr_t dest_addr,
-			uint64_t addr, uint64_t key,
-			enum fi_datatype datatype,
-			enum fi_op op, void *context,
-			uint64_t flags);
-
-ssize_t psmx2_atomic_compwrite_generic(
-			struct fid_ep *ep,
-			const void *buf,
-			size_t count, void *desc,
-			const void *compare, void *compare_desc,
-			void *result, void *result_desc,
-			fi_addr_t dest_addr,
-			uint64_t addr, uint64_t key,
-			enum fi_datatype datatype,
-			enum fi_op op, void *context,
-			uint64_t flags);
 
 #ifdef __cplusplus
 }
